Fixed strange disassociated file issues with charmedia stuff
