+++ BasicDb.Models/BasicDb.Models.csproj
@@ -58,11 +58,10 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CharacterModels.cs" />
+    <Compile Include="CharMediaModels.cs" />
     <Compile Include="ItemModels.cs" />
     <Compile Include="MediaModels.cs" />
     <Compile Include="CharItemModels.cs" />
-    <Compile Include="EditCharItem.cs" />
-    <Compile Include="PostCharItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
