--- BasicDb.Models/PostCharItem.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BasicDb.Models
-{
-    public class PostCharItem
-    {
-        [Required]
-        public int CharId { get; set; }
-        [Required]
-        public int ItemId { get; set; }
-    }
-}
