+++ BasicDb.Services/BasicDb.Services.csproj
@@ -57,15 +57,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-
     <Compile Include="CharacterService.cs" />
-
+    <Compile Include="CharMediaService.cs" />
     <Compile Include="ItemService.cs" />
-
     <Compile Include="MediaServices.cs" />
-
-    <Compile Include="CharItemService.cs"/>
-
+    <Compile Include="CharItemService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -74,7 +70,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\BasicDb.Data\BasicDb.Data.csproj">
-
       <Project>{562b542b-f014-457a-b16a-b9a055615c2b}</Project>
       <Name>BasicDb.Data</Name>
     </ProjectReference>
@@ -82,7 +77,6 @@
       <Project>{ef07a7ab-93df-4daa-b4a0-83502d308857}</Project>
       <Name>BasicDb.Models</Name>
     </ProjectReference>
-
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
