+++ BasicDb.Services/MediaServices.cs
@@ -24,7 +24,7 @@ namespace BasicDb.Services
                 {
                     AddedBy = _userId,
                     Name = media.Name,
-                    MediaType = media.MediaType,
+                    Medium = media.MediaType,
                     Description = media.Description,
                     CreatedOn = DateTime.Now,
                     ModifiedOn = DateTime.Now
@@ -43,7 +43,7 @@ namespace BasicDb.Services
             {
                 using (var ctx = new ApplicationDbContext())
                 {
-                    var query = ctx.Media.Select(e => new MediaGet { MediaId = e.MediaId, Title = e.Title, MediaType = e.MediaType, Description = e.Description, AddedBy = e.User.UserName });
+                    var query = ctx.Media.Select(e => new MediaGet { MediaId = e.MediaId, Title = e.Name, MediaType = e.Medium, Description = e.Description, AddedBy = e.User.UserName });
                     return query.ToArray();
                 }
             }
@@ -61,7 +61,7 @@ namespace BasicDb.Services
                     {
                         MediaId = entity.MediaId,
                         Title = entity.Name,
-                        MediaType = entity.MediaType,
+                        MediaType = entity.Medium,
                         Description = entity.Description,
                     };
                 }
@@ -78,7 +78,7 @@ namespace BasicDb.Services
                 var entity = ctx.Media.Single(e => e.MediaId == media.MediaId && e.AddedBy == _userId);
                 entity.MediaId = media.MediaId;
                 entity.Name = media.Title;
-                entity.MediaType = media.MediaType;
+                entity.Medium = media.MediaType;
                 entity.Description = media.Description;
                 entity.AddedBy = media.AddedBy;
                 //entity.ModifiedUtc = DateTimeOffset.UtcNow;
